uHAL clean up : Split 'Utilities.hpp' and 'Utilities.cpp' into 3 header files and 3 cpp files; refs issue #72
