+++ uhal/tests/src/common/test_log.cxx
@@ -36,9 +36,12 @@
 */
 
 
-#include "uhal/Utilities.hpp"
-#include "uhal/log/log.hpp"
 #include <iostream>
+
+#include "uhal/log/log.hpp"
+
+#include "uhal/utilities/files.hpp"
+
 
 using namespace uhal;
