+++ uhal/uhal/include/uhal/XmlParser.hpp
@@ -45,7 +45,6 @@
 
 #include "pugixml.hpp"
 #include "uhal/log/log.hpp"
-#include "uhal/Utilities.hpp"
 #include "uhal/log/exception.hpp"
 
 namespace uhal
