+++ uhal/uhal/include/uhal/utilities/bits.hpp
@@ -1,0 +1,58 @@
+/*
+---------------------------------------------------------------------------
+
+    This file is part of uHAL.
+
+    uHAL is a hardware access library and programming framework
+    originally developed for upgrades of the Level-1 trigger of the CMS
+    experiment at CERN.
+
+    uHAL is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    uHAL is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with uHAL.  If not, see <http://www.gnu.org/licenses/>.
+
+
+      Andrew Rose, Imperial College, London
+      email: awr01 <AT> imperial.ac.uk
+
+      Marc Magrans de Abril, CERN
+      email: marc.magrans.de.abril <AT> cern.ch
+
+      Tom Williams, Rutherford Appleton Laboratory, Oxfordshire
+      email: tom.williams <AT> cern.ch
+
+---------------------------------------------------------------------------
+*/
+
+
+#ifndef _uhal_utilities_bits_hpp_
+#define _uhal_utilities_bits_hpp_
+
+
+#include <stdint.h>
+
+
+namespace uhal
+{
+  namespace utilities
+  {
+    /**
+      Helper function to calculate the number of zero-bits at the righthand end of a 32-bit number
+      @param aValue a 32-bit number whose trailing zero-bits are to be counted
+      @return the number of trailing zero-bits
+    */
+    unsigned int TrailingRightBits ( uint32_t aValue );
+  }
+}
+
+
+#endif
