+++ uhal/uhal/include/uhal/utilities/files.hpp
@@ -27,17 +27,14 @@
       Marc Magrans de Abril, CERN
       email: marc.magrans.de.abril <AT> cern.ch
 
+      Tom Williams, Rutherford Appleton Laboratory, Oxfordshire
+      email: tom.williams <AT> cern.ch
+
 ---------------------------------------------------------------------------
 */
 
-/**
-	@file
-	@author Andrew W. Rose
-	@date 2012
-*/
-
-#ifndef _uhal_Utilities_hpp_
-#define _uhal_Utilities_hpp_
+#ifndef _uhal_utilities_files_hpp_
+#define _uhal_utilities_files_hpp_
 
 #include <vector>
 #include <string>
@@ -74,42 +71,12 @@
 
 #include "boost/unordered_map.hpp"
 
-// #ifdef __GNUC__
-// #include <ext/hash_map>
-// #else
-// #include <hash_map>
-// #endif
-//
-// // ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-// namespace std
-// {
-//   using namespace __gnu_cxx;
-// }
-//
-// //! Working in the __gnu_cxx namespace
-// namespace __gnu_cxx
-// {
-//   //! Add a hash function for a c++ std::string
-//   template<> struct hash< std::string >
-//   {
-//     //! implement the hash by calling the hash for the equivalent c-string
-//     size_t operator() ( const std::string& x ) const
-//     {
-//       using namespace uhal;
-//       return hash< const char* >() ( x.c_str() );
-//     }
-//   };
-// }
-// // ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
 
 // ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 namespace uhal
 {
   namespace exception
   {
-    //! Exception class to handle the case where the string will not fit into a 32-bit number.
-    UHAL_DEFINE_EXCEPTION_CLASS ( StringNumberWillNotFitInto32BitNumber , "Exception class to handle the case where the string will not fit into a 32-bit number." )
     //! Exception class to handle the case where the string is not a comma-delimiter list of URIs.
     UHAL_DEFINE_EXCEPTION_CLASS ( UriListParsingError , "Exception class to handle the case where the string is not a comma-delimiter list of URIs." )
     //! Exception class to handle the case where a URI contains a non-supported protocol.
@@ -409,350 +376,6 @@ namespace uhal
     }
 
 
-
-  }
-}
-// ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-
-// ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-namespace uhal
-{
-  namespace utilities
-  {
-    /**
-    	Helper function to make debugging failures when parsing XML files easier
-    	@param aLoadResult the result of the parsing
-    	@param aPath the full filename of file whose parsing failed
-    	@param aFile a byte vector containing the contents of the file (stored like this because the file could be either local or retrieved by HTTP)
-    */
-    void PugiXMLParseResultPrettifier ( const pugi::xml_parse_result& aLoadResult , const boost::filesystem::path& aPath , const std::vector<uint8_t>& aFile );
-  }
-}
-// ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-
-// ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-namespace uhal
-{
-  namespace utilities
-  {
-    /**
-    	Helper function to calculate the number of zero-bits at the righthand end of a 32-bit number
-    	@param aValue a 32-bit number whose trailing zero-bits are to be counted
-    	@return the number of trailing zero-bits
-    */
-    unsigned int TrailingRightBits ( uint32_t aValue );
-  }
-}
-// ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-
-// ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-namespace uhal
-{
-  namespace utilities
-  {
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , std::string& aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        aTarget = lAttr.value();
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
-
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , const char* aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        aTarget = lAttr.value();
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
-
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , int32_t& aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        std::string lAttrStr ( lAttr.value() );
-        std::stringstream ss;
-
-        //if string is of the form "x89abcdef" , "X89abcdef" , "0x89abcdef" , "0X89abcdef"
-        if ( lAttrStr.size() > 2 )
-        {
-          if ( ( lAttrStr[1] == 'x' ) || ( lAttrStr[1] == 'X' ) )
-          {
-            ss << std::hex << lAttrStr.substr ( 2 );
-          }
-          else
-          {
-            ss << lAttrStr;
-          }
-        }
-        else if ( lAttrStr.size() > 1 )
-        {
-          if ( ( lAttrStr[0] == 'x' ) || ( lAttrStr[0] == 'X' ) )
-          {
-            ss << std::hex << lAttrStr.substr ( 1 );
-          }
-          else
-          {
-            ss << lAttrStr;
-          }
-        }
-        else
-        {
-          ss << lAttrStr;
-        }
-
-        // ss >> aTarget;
-        // aTarget = lAttr.as_int();
-
-        if ( ss.str().size() > 10 )
-        {
-          exception::StringNumberWillNotFitInto32BitNumber lExc;
-          log ( lExc , "XML attribute " , Quote ( aAttrName ) , " has value " , Quote ( ss.str() ) , " which is too big to fit into 32-bit number" );
-          throw lExc;
-        }
-
-        int64_t lTarget;
-        ss >> lTarget;
-
-        if ( lTarget>>32 )
-        {
-          exception::StringNumberWillNotFitInto32BitNumber lExc;
-          log ( lExc , "XML attribute " , Quote ( aAttrName ) , " has value " , Quote ( ss.str() ) , " which is too big to fit into 32-bit number" );
-          throw lExc;
-        }
-
-        aTarget = ( int32_t ) ( lTarget );
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , uint32_t& aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        std::string lAttrStr ( lAttr.value() );
-        std::stringstream ss;
-
-        //if string is of the form "x89abcdef" , "X89abcdef" , "0x89abcdef" , "0X89abcdef"
-        if ( lAttrStr.size() > 2 )
-        {
-          if ( ( lAttrStr[1] == 'x' ) || ( lAttrStr[1] == 'X' ) )
-          {
-            ss << std::hex << lAttrStr.substr ( 2 );
-          }
-          else
-          {
-            ss << lAttrStr;
-          }
-        }
-        else if ( lAttrStr.size() > 1 )
-        {
-          if ( ( lAttrStr[0] == 'x' ) || ( lAttrStr[0] == 'X' ) )
-          {
-            ss << std::hex << lAttrStr.substr ( 1 );
-          }
-          else
-          {
-            ss << lAttrStr;
-          }
-        }
-        else
-        {
-          ss << lAttrStr;
-        }
-
-        // ss >> aTarget;
-        // aTarget = lAttr.as_uint();
-
-        if ( ss.str().size() > 10 )
-        {
-          exception::StringNumberWillNotFitInto32BitNumber lExc;
-          log ( lExc , "XML attribute " , Quote ( aAttrName ) , " has value " , Quote ( ss.str() ) , " which is too big to fit into 32-bit number" );
-          throw lExc;
-        }
-
-        uint64_t lTarget;
-        ss >> lTarget;
-
-        if ( lTarget>>32 )
-        {
-          exception::StringNumberWillNotFitInto32BitNumber lExc;
-          log ( lExc , "XML attribute " , Quote ( aAttrName ) , " has value " , Quote ( ss.str() ) , " which is too big to fit into 32-bit number" );
-          throw lExc;
-        }
-
-        aTarget = ( uint32_t ) ( lTarget );
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , double& aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        aTarget = lAttr.as_double();
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , float& aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        aTarget = lAttr.as_float();
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
-
-    /**
-    	Helper function to retrieve a named attribute from a PugiXML node and cast it to the correct type
-    	@param aNode a node from which the attribute is to be extracted
-    	@param aAttrName the name of the attribute to be extracted
-    	@param aTarget a variable into which the attribute's value id to be written
-    	@return success/failure status
-    */
-    template < bool DebugInfo >
-    bool GetXMLattribute ( const pugi::xml_node& aNode , const char* aAttrName , bool& aTarget )
-    {
-      pugi::xml_attribute lAttr = aNode.attribute ( aAttrName );
-
-      if ( ! lAttr.empty() )
-      {
-        aTarget = lAttr.as_bool();
-        return true;
-      }
-      else
-      {
-        if ( DebugInfo )
-        {
-          log ( Error() , "Failed to get attribute " , Quote ( aAttrName ) , " from XML node." );
-        }
-
-        return false;
-      }
-    }
 
   }
 }
